crud

+++ src/main/java/invoice/com/demo/features/quotation/dto/QuotationItemUpdateRequest.java
@@ -2,10 +2,12 @@ package invoice.com.demo.features.quotation.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class QuotationItemUpdateRequest {
     private Long id;
     private Long productId;
     private Integer quantity;
-    private Double unitPrice;
+    private BigDecimal unitPrice;
 }
